+++ SpriterDotNet.Unity/Assets/SpriterDotNet/SpriterImporter.cs
@@ -19,8 +19,9 @@ namespace SpriterDotNetUnity
     {
         private static readonly string AutosaveExtension = ".autosave.scml";
         private static readonly string[] ScmlExtensions = new string[] { ".scml" };
+        private static readonly string ObjectNameSprites = "Sprites";
+        private static readonly string ObjectNameMetadata = "Metadata";
 
-        public static float DeltaZ = -0.001f;
         public static bool UseNativeTags = true;
 
         public static event Action<SpriterEntity, GameObject> EntityImported = (e, p) => { };
@@ -51,8 +52,8 @@ namespace SpriterDotNetUnity
             foreach (SpriterEntity entity in spriter.Entities)
             {
                 GameObject go = new GameObject(entity.Name);
-                GameObject sprites = new GameObject("Sprites");
-                GameObject metadata = new GameObject("Metadata");
+                GameObject sprites = new GameObject(ObjectNameSprites);
+                GameObject metadata = new GameObject(ObjectNameMetadata);
 
                 SpriterDotNetBehaviour behaviour = go.AddComponent<SpriterDotNetBehaviour>();
                 behaviour.UseNativeTags = UseNativeTags;
@@ -94,15 +95,45 @@ namespace SpriterDotNetUnity
         private static GameObject CreatePrefab(GameObject go, string folder)
         {
             string prefabPath = folder + "/" + go.name + ".prefab";
-            GameObject existing = AssetDatabase.LoadAssetAtPath<GameObject>(prefabPath);
+            GameObject existingPrefab = AssetDatabase.LoadAssetAtPath<GameObject>(prefabPath);
 
             GameObject prefab;
-            if (existing != null) prefab = PrefabUtility.ReplacePrefab(go, existing, ReplacePrefabOptions.Default);
+            if (existingPrefab != null) prefab = ReplacePrefab(go, existingPrefab, prefabPath);
             else prefab = PrefabUtility.CreatePrefab(prefabPath, go, ReplacePrefabOptions.Default);
 
             GameObject.DestroyImmediate(go);
 
             return prefab;
+        }
+
+        private static GameObject ReplacePrefab(GameObject go, GameObject prefab, string path)
+        {
+            GameObject existing = GameObject.Instantiate(prefab);
+            MoveChild(go, existing, ObjectNameSprites);
+            MoveChild(go, existing, ObjectNameMetadata);
+
+            SpriterDotNetBehaviour sdnbNew = go.GetComponent<SpriterDotNetBehaviour>();
+            SpriterDotNetBehaviour sdnbExisting = existing.GetComponent<SpriterDotNetBehaviour>();
+            sdnbExisting.ChildData = sdnbNew.ChildData;
+            sdnbExisting.EntityIndex = sdnbNew.EntityIndex;
+            sdnbExisting.SpriterData = sdnbNew.SpriterData;
+            sdnbExisting.UseNativeTags = sdnbNew.UseNativeTags;
+
+            GameObject createdPrefab = PrefabUtility.ReplacePrefab(existing, prefab, ReplacePrefabOptions.Default);
+            GameObject.DestroyImmediate(existing);
+            return createdPrefab;
+        }
+
+        private static void MoveChild(GameObject from, GameObject to, string name)
+        {
+            Transform toChild = to.transform.Find(name);
+            GameObject.DestroyImmediate(toChild.gameObject);
+
+            Transform fromChild = from.transform.Find(name);
+            fromChild.SetParent(to.transform);
+            fromChild.localPosition = Vector3.zero;
+            fromChild.localRotation = Quaternion.identity;
+            fromChild.localScale = Vector3.one;
         }
 
         private static void CreateSprites(SpriterEntity entity, ChildData cd, Spriter spriter, GameObject parent)
@@ -123,7 +154,7 @@ namespace SpriterDotNetUnity
                 cd.SpritePivots[i] = pivot;
                 cd.Sprites[i] = child;
 
-                child.transform.localPosition = new Vector3(0, 0, DeltaZ * i);
+                child.transform.localPosition = Vector3.zero;
 
                 child.AddComponent<SpriteRenderer>();
             }