+++ SpriterDotNet/FrameData.cs
@@ -10,14 +10,21 @@ namespace SpriterDotNet
     public class FrameData
     {
         public List<SpriterObject> SpriteData { get; private set; }
-        public IDictionary<string, SpriterObject> PointData { get; private set; }
-        public IDictionary<int, SpriterObject> BoxData { get; private set; }
+        public Dictionary<string, SpriterObject> PointData { get; private set; }
+        public Dictionary<int, SpriterObject> BoxData { get; private set; }
 
         public FrameData()
         {
             SpriteData = new List<SpriterObject>();
             PointData = new Dictionary<string, SpriterObject>();
             BoxData = new Dictionary<int, SpriterObject>();
+        }
+
+        public void Clear()
+        {
+            SpriterObjectPool.ReturnChildren(SpriteData);
+            SpriterObjectPool.ReturnChildren(PointData);
+            SpriterObjectPool.ReturnChildren(BoxData);
         }
     }
 }